requirements.txt
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.4
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Session==0.8.0
Flask-Testing==0.8.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.2.2
pytest-flask==1.3.0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.7
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
zope.interface==6.1
